// src/wl/types.ts
export type Data = {
  id: number;
  pid: number;
  container_num: string;
  container_position: string;
  logcode: string;
  sub_logcode: string;
  sub_type: number;
  member_id: number;
  member_name: string;
  mark_id: number;
  mark_name: string;
  receive_time: number;
  from_name: string;
  from_phone: string;
  from_address: string;
  to_name: string;
  to_phone: string;
  to_address: string;
  agent_manager_id: number;
  weight: string;
  chargeable_weight: string;
  weight_record: string;
  volume: string;
  chargeable_volume: string;
  volume_record: string;
  package_record: string;
  tariff_record: string;
  quoted_record: string;
  forklift_record: string;
  pick_record: string;
  luck_record: string;
  delivery_record: string;
  loading_record: string;
  advancepayment_record: string;
  crane_record: string;
  other_record: string;
  unit_price: string;
  weight_limit: number;
  sub_total: string;
  total: string;
  writeoff_price: string;
  package_price: string;
  forklift_price: string;
  crane_price: string;
  loading_price: string;
  pick_price: string;
  quoted_price: string;
  tariff_price: string;
  luck_price: string;
  delivery_price: string;
  other_price: string;
  discount_price: string;
  advancepayment_price: string;
  net_weight: string;
  transport_id: string;
  transport_name: string;
  goods_name: string;
  goods_number: number;
  goods_number_start: number;
  goods_number_end: number;
  containing_numder: number;
  position: string;
  material: string;
  purpose: string;
  warehousing_pic: string;
  usercomments: string;
  warehousingremarks: string;
  expresstracking: string;
  deliveryway: string;
  situation: string;
  spot_code: string;
  status: number;
  drawer_manager_id: number;
  drawer_manager_name: string;
  delivery_manager_id: number;
  delivery_manager_name: string;
  signingtime: number;
  voucherpic: string;
  order_numder_details?: any;
  payment_method_id: string;
  payment_method_name: string;
  payment_status: number;
  is_cancellation: number;
  is_split: number;
  pack_time: number;
  is_make_bill: number;
  updatetime: number;
  addtime: string;
  sub_order: string;
  haveChild: boolean;
  desc: string;
};